+++ NEWS.md
@@ -1,5 +1,8 @@
 # tune 0.0.0.9002
 
+* New `fit_resamples()` for fitting a set of resamples that don't require any
+  tuning.
+
 * Changed `summarise.tune_results()` back to `estimate.tune_results()`
 
 # tune 0.0.0.9000